Compile js http cache only for wasm target
