+++ cloudflare_worker/src/lib.rs
@@ -107,7 +107,7 @@ pub async fn create_signed_exchange(
         .map_err(to_js_error)?;
     let signer = ::sxg_rs::signature::js_signer::JsSigner::from_raw_signer(signer);
     let subresource_fetcher = sxg_rs::fetcher::js_fetcher::JsFetcher::new(subresource_fetcher);
-    let mut header_integrity_cache = sxg_rs::http_cache::JsHttpCache {
+    let mut header_integrity_cache = sxg_rs::http_cache::js_http_cache::JsHttpCache {
         get: header_integrity_get,
         put: header_integrity_put,
     };
